+++ aiotdlib/api/__init__.py
@@ -1030,6 +1030,9 @@ BaseObject._all = {
 
 __all__ = [
     'AioTDLibError',
+    'BadRequest',
+    'NotFound',
+    'Unauthorized',
     'API',
     'BaseObject',