+++ aiotdlib/api/errors/__init__.py
@@ -1,5 +1,6 @@
 from .error import (
     AioTDLibError,
     BadRequest,
+    NotFound,
     Unauthorized,
 )